requirements.txt
flask
flask-smorest
flask_sqlalchemy
pymysql
marshmallow
python-dotenv
requests 